reintroduced legacy methods to maintain backward compatibility
